Integration tests renamed
